Move more plotting dependend tests
