+++ tests/filters/test_filters.py
@@ -8,7 +8,8 @@ import pytest
 from vtk import VTK_QUADRATIC_HEXAHEDRON, VTK_QUADRATIC_TRIANGLE
 
 import pyvista
-from pyvista import _vtk, examples
+from pyvista import examples
+from pyvista.core import _vtk_core
 from pyvista.core.errors import NotAllTrianglesError, VTKVersionError
 from pyvista.errors import MissingDataError
 
@@ -848,7 +849,7 @@ def test_glyph_cell_point_data(sphere):
 
 
 class InterrogateVTKGlyph3D:
-    def __init__(self, alg: _vtk.vtkGlyph3D):
+    def __init__(self, alg: _vtk_core.vtkGlyph3D):
         self.alg = alg
 
     @property
@@ -1159,7 +1160,7 @@ def test_resample():
 
 @pytest.mark.parametrize('use_points', [True, False])
 @pytest.mark.parametrize('categorical', [True, False])
-@pytest.mark.parametrize('locator', [None, _vtk.vtkStaticCellLocator()])
+@pytest.mark.parametrize('locator', [None, _vtk_core.vtkStaticCellLocator()])
 def test_probe(categorical, use_points, locator):
     mesh = pyvista.Sphere(center=(4.5, 4.5, 4.5), radius=4.5)
     data_to_probe = examples.load_uniform()
